ai_service/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0
pydantic-settings>=2.5.0
google-genai>=0.8.0
pinecone>=5.4.2
pypdf>=5.1.0
python-docx>=1.1.0
python-multipart>=0.0.12
python-dotenv>=1.0.1
httpx>=0.27.2
aiofiles>=24.1.0
numpy>=1.26.4
scikit-learn>=1.5.2
motor>=3.6.0
pymongo>=4.10.1
tiktoken>=0.8.0
langchain>=0.3.7
langchain-google-genai>=2.0.5
langchain-pinecone>=0.2.0
langchain-core>=0.3.15
datasets>=3.1.0
ragas>=0.2.6
